requirements.txt
tweepy==4.14.0
python-dotenv
pandas
pillow
requests

